requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
arrow==1.3.0
attrs==23.1.0
cachetools==5.3.2
cassiopeia==5.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
datapipelines==1.0.7
discord==2.3.2
discord.py==2.3.2
dnspython==2.4.2
frozenlist==1.4.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gspread==5.12.3
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
merakicommons==1.0.10
multidict==6.0.4
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pluggy==1.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
types-python-dateutil==2.8.19.14
tzdata==2023.3
urllib3==2.1.0
yarl==1.9.4